refactor(core): split large nodes.py into modular node files - Moved node functions into separate files under core/nodes:   - conversation.py   - domain_analysis.py   - context_retrieval.py   - query_classification.py   - command_generation.py   - information_generation.py   - tool_execution.py   - result_preparation.py   - display_result.py
