+++ src/os_assistant/core/nodes/nodes.py
@@ -1,0 +1,57 @@
+from __future__ import annotations
+
+from datetime import datetime
+from typing import TYPE_CHECKING
+
+from langchain.schema import HumanMessage, SystemMessage
+from langchain_ollama import ChatOllama
+from tracer.config import LogDomain
+
+from os_assistant.utils.settings import (
+    ASSISTANT_MODE,
+    DOMAINS,
+    MODEL_BASE_URL,
+    MODEL_NAME,
+)
+from os_assistant.utils.model_factory import model
+
+from os_assistant.parsers.setup import (
+    code_execute_parser,
+    command_response_parser,
+    domain_analysis_parser,
+    fixed_command_response_parser,
+    fixed_domain_analysis_parser,
+    fixed_info_response_parser,
+    fixed_query_type_parser,
+    info_response_parser,
+    parse_with_fix_and_extract,
+    query_type_parser,
+)
+from os_assistant.prompts.prompt_loader import load_prompt
+from os_assistant.pydantic_models.schemas import (
+    CommandResponse,
+    DomainAnalysis,
+    FinalResult,
+    InformationResponse,
+    QueryTypeResult,
+)
+from os_assistant.tools.agentic_rag.application.search import search_logs
+from os_assistant.tools.code_agent.wrapper import code_execute_tool
+from os_assistant.core.nodes.helpers import (
+    is_rag_enabled,
+    is_code_execution_enabled,
+    get_mode_description
+)
+
+if TYPE_CHECKING:
+    from os_assistant.core.state import AssistantState
+
+
+
+
+
+
+
+
+
+
