+++ src/os_assistant/core/nodes/registry.py
